+++ lib/api/isoexpress/api_isone_zonal_demand.dart
@@ -1,13 +1,21 @@
 library api.isone_zonal_demand;
 
 import 'dart:async';
+import 'dart:convert';
 import 'package:mongo_dart/mongo_dart.dart';
+import 'package:shelf/shelf.dart';
+import 'package:shelf_router/shelf_router.dart';
 import 'package:timezone/timezone.dart';
 import 'package:intl/intl.dart';
 import 'package:date/date.dart';
 
-// @ApiClass(name: 'zonal_demand', version: 'v1')
 class ZonalDemand {
+  ZonalDemand(this.db) {
+    coll = db.collection(collectionName);
+    _location = getLocation('America/New_York');
+  }
+
+  late Db db;
   late DbCollection coll;
   late Location _location;
   final DateFormat fmt = DateFormat("yyyy-MM-ddTHH:00:00.000-ZZZZ");
@@ -22,23 +30,43 @@ class ZonalDemand {
     'RI',
     'SEMA',
     'WCMA',
-    'NEMA'
+    'NEMA',
   ];
 
-  ZonalDemand(Db db) {
-    coll = db.collection(collectionName);
-    _location = getLocation('America/New_York');
+  final headers = {
+    'Content-Type': 'application/json',
+  };
+
+  Router get router {
+    final router = Router();
+
+    /// Return a list of Map elements like this:
+    /// ```dart
+    /// {
+    ///   'hourBeginning': '2017-01-01 00:00:00.000-0500',
+    ///   'DA_Demand': 11167.1,
+    ///   'RT_Demand': 11810.35,
+    ///   'DryBulb': 37,
+    ///   'DewPoint': 32,
+    /// }
+    /// ```
+    router.get('/zone/<zone>/start/<start>/end/<end>', (Request request,
+        String zone, String start, String end) async {
+      var aux = await apiGetZonalDemand(zone, start, end);
+      return Response.ok(json.encode(aux), headers: headers);
+    });
+
+    return router;
   }
 
-  /// http://localhost:8080/zonal_demand/v1/zone/isone/start/20170101/end/20170101
-  // @ApiMethod(path: 'zone/{zone}/start/{start}/end/{end}')
-  Future<List<Map<String, String>>> apiGetZonalDemand(
-      String zone, String start, String end) async {
-    var startDate = Date.parse(start);
-    var endDate = Date.parse(end);
 
-    Stream aux = _getHourlyData(zone, startDate: startDate, endDate: endDate);
-    var out = [];
+  /// http://localhost:8080/zonal_demand/v1/zone/isone/start/20170101/end/20170101
+  ///
+  Future<List<Map<String, dynamic>>> apiGetZonalDemand(
+      String zone, String start, String end) async {
+
+    var aux = await _getHourlyData(zone, startDate: start, endDate: end);
+    var out = <Map<String,dynamic>>[];
     var keys = [
       'hourBeginning',
       'DA_Demand',
@@ -46,7 +74,7 @@ class ZonalDemand {
       'DryBulb',
       'DewPoint'
     ];
-    await for (var e in aux) {
+    for (var e in aux) {
       /// each element is a list of 24 hours, prices
       for (var i = 0; i < e['hourBeginning'].length; i++) {
         out.add(Map.fromIterables(keys, [
@@ -58,27 +86,28 @@ class ZonalDemand {
         ]));
       }
     }
-    return out as FutureOr<List<Map<String, String>>>;
+    return out;
   }
 
 
-  /// Workhorse to extract the data ...
-  /// returns one element for each day
-  Stream _getHourlyData(String zone, {Date? startDate, Date? endDate}) {
-    var pipeline = [];
-    var match = {};
-    zone = zone.toUpperCase();
-    if (_canonicalZones.contains(zone)) match['zoneName'] = zone;
-
-    var date = {};
-    if (startDate != null) date['\$gte'] = startDate.toString();
-    if (endDate != null) date['\$lt'] = endDate.add(1).toString();
-    if (date.isNotEmpty) match['date'] = date;
-
-    pipeline.add({'\$match': match});
-    pipeline.add({
-      '\$project': {'_id': 0}
-    });
-    return coll.aggregateToStream(pipeline as List<Map<String, Object>>);
+  /// Return one element for each day
+  Future<List<Map<String,dynamic>>> _getHourlyData(String zone, {required String startDate, required String endDate}) {
+    var pipeline = <Map<String,Object>>[
+      {
+        '\$match': {
+          'zoneName': zone.toUpperCase(),
+          'date': {
+            '\$lte': Date.parse(endDate).toString(),
+            '\$gte': Date.parse(startDate).toString(),
+          },
+        },
+      },
+      {
+        '\$project': {
+          '_id': 0,
+        }
+      }
+    ];
+    return coll.aggregateToStream(pipeline).toList();
   }
 }